drakrun: Rename vmconf as vm
